fix(ast_printer.c): print mutable variable declarations as 'let-mut'
When printing a 'let mut' declaration, the AST printer produces '(let
mut var value)' which can be confusing since immutable declarations also
start with a 'let' follwing the variable name.
Change it to print 'let-mut' instead so that the second word is always
the variable name in both the type of declarations.