fix(parser): beware of EOF while parsing arguments
When parsing arguments to a function call, lets add a check to see if we
have reached EOF before attempting to consume any tokens. This fixes the
bug where argument parser failed to report an unclosed parenthesis
becuase we attempted to parse an expression regardless of an EOF.

This also gives us while loop condition that reads much nicer than the
original `for (;;)`.