refactor(parser): move list append logic to a helper
When parsing a block, we were appending stmts to a list of nodes,
inline. Move this logic to helper 'stmt_list_append' which takes in
head, tail and node to append.

This helper can also be used to replace the logic in insert_stmt, which
is also just a funtion that appends to linked list.