fix(tests): dont expect statments at top-level
So far, our tests have been expecting the top level code to contain the
statements directly. We recently introduced items, which will be the
only nodes that stay at top-level. Statements live nested inside
functions.

Update our tests to reflect this. We achive this by writing a helper
`stmt2sexper` which builds on top of code2sexpr and wraps the statement
code to a function before passing that to onec and unwraps the
statment's S-expressions. This allows allows us to keep the existing
tests.

While we are at it, lets also add a test for functions.