feat: revamp argument parsing
Write a argument parsing routine that is more extensible than what we
already have. We are expecting the compiler to have several flags in the
future, lets prepare for it.

Currently, the compiler does not execute or compile any code and only
transpiles code into an S-expression. While this is a nice feature the
compiler should have, it should not be the default behaviour.
Add a --s-expr (-S) flag that requests this behaviour explicitly.
Since we do not have infrastructure for compiling/executing code yet, we
error out with a "NOT IMPLEMENTED" message in the code paths that would
normally perform code compilation/execution.

Since t/lib.py relied on this behaviour of printing s-expressions by
default, it must now explicitly request it.