feat(vec): add vec2: a nicer implementation of vec
Although we already have a dynamic array implementation called 'vec',
its API is not type safe: Each read-write operation requires knowing the
type of element the vec contains. These routines do not complain if an
incorrect type is used.

Lets introduce a new vec2 API which addresses these issues. vec2
routines are directly implemented as macros, unlike the original vec
which made use of intermediate function which were unaware of the
element type.

The VEC2 macro creates a vec struct specific to an element type, and the
macros vec2_push and vec2_get operate on these structs. These only take
a pointer to a vec2 struct and an element/index respectively, specifying
types is not required for reading and writing.
Since these read and write from a type-aware struct, the r/w macros will
fail if incorrect types are used.

This commit only adds this API, we will gradually want migrate uses of
old vec API to this.