feat(ast): add callback context to walkers
When walking an AST, a walker might need to persist additional context.
For instance, a type-checker might have a symbol table to keep track of
the types of various symbols.

Add an additional 'ctx' fields to all the AST walker macros. These
macros now take a type which defines the type the 'ctx' field should be.

Since the AST printer does not need any context, let it use an empty
struct as context, which we intentionally don't initialize.