feat: add 'unit' type
Add a 'unit' typename to represent return type of functions that don't
return anything. There is currently no syntax to specify this
explicitly, instead user may leave the return type unspecified to make
the return type unit.

Teach AST parser about this and make it output a representation of unit
type.
Lets use '()' for this. This is like an 'empty tuple' that essentially
represents 'nothing'. We do not have a tuple type in the language yet,
however we do plan to add them. Using an empty tuple to represent lack
of a value therefore makes sense.

Now since our functions and lambdas that don't have a return type will
cause the AST printer to print ':()' after their parameters, update the
tests to reflect this. In particular, lib.py assumes a very specific
form of S-expression of the 'main' function and relies on it to extract
the statements from it. Make it aware of our new addition.