fix(vec): incorrect memcpy size in push
when pushing an element to vec, we were copying the item with memcpy
with size set to sizeof(item). This is incorrect because item is
expected to be a pointer and not the item itself. Correct it to
sizeof(*item), which gives us the size of the actual type we are
storing.