feat(test): add 'invoke_onec' helper
We have functions for converting code and statements to S-expressions,
however they only provide us with the stdout of onec and thus cannot be
used to make assertions about exit code of onec.

Lets add a lower-level 'invoke_onec' helper that allows specifying args
for onec directly and also returns the exit code along with the stdout.