fix(parser): correctly parse single-token and empty programs
We were using a do-while loop in the parser to parse and insert items to
AST until the *lexer* reaches EOF. This is wrong because the program may
be empty and the lexer will threfore produce only a single EOF token. In
this case the parser would still continue parsing looking for an item,
when there is none and report a syntax error. This shouldn't happen
because an empty program is still a valid program.

We fix this by swapping out the do-while loop with a while loop.

However, using lexer's EOF state is a problem when a (malformed) program with
only a single character is input: the first advance causes the lexer to
reach EOF, and the while loop skips over without looping even once, the
first token that we advanced never gets processed.

We fix this by using *parser's* EOF state as the loop condition.
Add a 'is_eof' helper to help us determine if parser has reached end of
token stream.