refactor(parser): don't rely on lexer's EOF state
Since lexer's EOF state might not always determine if parser has
actually consumed all the tokens, it might be not be a good idea to rely
on it. We instead use parser's own EOF state which reliably tells if we
have finished processing all the tokens by checking if the current token
is an EOF.

Since we don't want to use 'lex_eof' anywhere else other than the lexer
iteself, lets hide it from the rest of the world by removing its
declaration from lex.h and making it static.