feat(parser): add unit type syntax
Unit types represent "no value", simmilar to C's void type.
A function returning nothing has a return type of unit.

We plan to implement unit types as zero-sized tuples, however we do not
have syntax for tuples yet. Lets write only enough of the tuple parsing
logic to be sufficient for parsing unit types.

Add a test for this new type-name to ensure we are able to parse it.