refactor(tests): remove redundant line squishing
In an early stage, onec was not capable of reading multi-line code. In
'test_block.py', we had a function which would take a multi-line source
code and squish it down to a single line for onec to interpret
correctly.

Since we no longer have that limitation now, get rid of it.