fix(typecheck): 'TYPE_NAME_*' -> 'TYPE_RES_*'
In typecheck.c, we were using ast_typename_type variants for assignemnt
and comparision with 'typeres_type' enums. This is wrong but somehow our
tests don't detect this mistake. This is likely have to do with the fact
that all 'typeres_type' has all the coresponding variants from
'ast_typename_type' in the exact order of the former. This does not make
it any less wrong however: if the order of variants in 'typeres_type'
change or new variants are added, it will cause failures everywhere
throught the typecheker.

Replace these incorrect enum variant uses with the correct variants from
'typeres_type' enum.