feat: introduce fixed-sized integer types
Simillar to Rust, lets have integer types of different size and sign:

    u8, i8, u16, i16, u32, i32

The prefix tells if its a signed or unsized type and the number tells
the size in bits.

Get rid of the number type we had, which did not specify any of these
details.

Add parser rule and AST nodes for this.
Teach AST printer and typechecker about this new node.

Run a find & replace in our tests to change all the 'number' instances
to 'i32'.

Differ typechecking to verify if the assignment type is compatible for
future commit.

Floating point numbers can be implemented in future with simmilar
syntax, i.e. as f32 and f64.