fix(typecheck): deep-clone tuple typeres in typeres_dup
We were not duplicating inner contents of tuple typeres in typeres_dup,
this caused assignments of tuple type to be incorrectly reported as
incompatible. Fix this by duplicating all the inner items of the tuple.

Since duplicating both function and tuple type involves deep-copying a
vec_typeres, lets extract that logic into a common function.

Add a simple test that would fail without this patch.