fix(typecheck): duplicate left oprand's typeres for binary operations
When typechecking binary operator were were only copying the base type
of the left operand. This works right now because we don't expect number
and string nodes to have any other data than the type tag. However, that
will change when integer types of different sizes get introduced.
Duplicate the entire type resolution to resolve this.
Do this only for binary operators that operate on integers.