feat(typecheck): implement typechecking for integer types
Since we now have AST nodes and syntax for integer types: u8, u16, u32,
i8, i16 and i32, teach typechecker how to validate their uses.

Specifically, want the program to complain if an integer with different
size and/or sign is used in a context where a differently signed/sized
integer is expected.

One interesting case is integer literals, since there is no way (yet) to
write integer literals while specifying the specific type the literal is
of, we must assume i32 as the type untill we get to a node/context where
we can determine what is really expected. We achive this by deferring
a concrete type resolution until such context is discovered (via
default_until_inferred flag). If we are never able to determine the
concrete type, we assume i32 as the final type.