refact(Makefile): list objects intead of sources
In the Makefile, we were listing our source files as:

	SRCS += alloc.c
	SRCS += ast.c
	..

Instead of this, let's build up a list of _objects_ that will be compiled
from our sources like:

	OBJ += alloc.o
	OBJ += ast.o
	..

and figure out the source file based on the object name instead of the
other way around.

This makes the Makefile consistently follow the idea of "specify build
targets, infer sources".