refact(Makefile): separate lib and onec objects
Separate onec library objects from onec executable objects.
Library objects are those which provide a public interface and can be
tested separately. 'main.c' does not fall under this category, so let's
separate it from the lib objects.

This leaves us with two different lists of objects:

* LIB_OBJ   Everything required to build up something like a "libonec",
            although that is not a thing yet.

* ONEC_OBJ  Everything required to build up the entire onec executable.
	    This includes objects in LIB_OBJ.

Although there is a separation in naming, all of the objects are
compiled with the same recipe.