feat(lex): track line and col number of tokens
When reporting syntax errors, it is often helpful to the user to be able
to see the line/column of the source code that triggered it.
We were unable to do this because our lexer did not produce information
about line and col numbers of the tokens it emitted.

Add 'line' and 'col' fields in the token struct that tell us the
beginning of the token in the source file.
Teach lexer how to fill up this information.