feat(tests/lib): add a python wrapper over code2token-list
In previous two commits, we introduced a program 'code2token-list' that
invokes lexer to lex a program and outputs tokens to stdout.

To be able to use that in pytest test cases, we need a wrapper that adds
a python interface to this program. Add function 'code2token_list' in
't/lib.py' to do exactly that.