fix(lex): consume at most one decimal point in a number
When tokenizing numbers, we did not consider that a malformed program
might have multiple decimal points and continued lexing with any number
of '.' we see.

Fix this by setting a flag when we see a decimal point and stop lexing
when we see another.

Add a test to verify this behaviour.