refact(lex): hold current token start in lexer
Remember the position of current token we are lexing. This includes a
pointer to the character where the current token starts and line and
column of that character in the source.

Since now lexer is aware of the position of start character the token it
should make next, we can update make_token to only take lexer as
parameter and it can figure out the span, line and column of the token.

This will help us further refactor lexer to use functions such as
'peek', 'next/advance' and 'match' instead of doing those operations
inline.

Note that here by "current" token I really mean the token lexer is
currently working on. In other words, it is the _next_ token that will
be emitted by the lexer if everything goes right.