refact(lex): add util 'match'
Add untility function 'match', similar to one in parser, which matches
the given character against given one and advances the lexer head if the
match succeeds, returning true. If the character does not match, returns
false.

This helps us generalize a specific use of peek() and advance().

Something like this:

	if (peek(..) == expected) {
		advance(..);
		// do something
	}

Can be reduced to:

	if (match(.., expected)) {
		// do something
	}