fix(parser): incorrect handling of single char escapes
When parsing escape sequences in a string, we were not advancing our
head after we have handled a single character escape such as '\n'.

This lead to wierd behaviour where those escaped produced two
characters: 1. the result of the escape and 2. the escape character
itself (eg. 'n' in case of '\n').

Fix this by advancing the head when we process those escapes.

Add a basic test to verify this behaviour.