feat: allow escaping a quote (") in strings
Most languages allow escaping the characters used for enclosing strings.
We did not have such an escape sequence. Let's add '\"' which escapes
the '"' character without terminating the string.

Since we handle escapes sequences in parser, lexer has no awareness of
it. However, in order to implement this, we need _some_ kind of escape
awareness in the lexer to be able to determine if the '"' was escaped or
marked the end of the string.

Reimplementing entire escape parsing in lex is not desired, since we
don't need _all_ of that functionality. Let's implement just enough of
this to be able to skip single character escapes such as '\"'.

Note that this implementation is not aware of multi-character escapes
such as \xNN and \NN, it only skips checking if the character right next
to a backward slash is '"' to detect string termination. This works for
detecting correct end of string but does not give lexer full awareness
of the escape sytanx, and does not even try to validate it's
correctness -- only parser can do that.

Add lexer and parser tests to verify this is functional.