fix(lexer): lex_advance: incorrect declaration in header
In lexer.c, we have 'lex_advance(lexer_t*)' taking a pointer to the
lexer_t struct. However, in lexer.h, its declaration did not have that
parameter.

Moreover, in parser, we were calling the said function as:

    lex_advance(parser->lexer)

which passes not a pointer to the lexer, but the lexer itself. It should
instead read:

    lex_advance(&parser->lexer)

Do it.

For some reason this worked on the machine I have been so far working
on and went unnoticed. On my current machine however, with gcc 13.1.1,
this resulted in "stack smashing".