fix(ast): add 'free_ast' for freeing entire ast
We have 'free_ast_item' which frees 'ast_item_node' type. However, this
struct is also used as a linked list. When there are multiple items in
the AST, the 'next' field on this struct points to the next item node.
We were not walking this linked list to free everything in the AST.
This leaked memory if the AST had more than one item nodes (i.e.
function nodes -- the only items we have for now.)

However, we don't want to perform the linked list traversal in
'free_ast_item', as the name of that function suggests that it is only
responsible for freeing a single item node.

Let's instead add another function 'free_ast' which walks the linked
list of item nodes and delegates the free operation for each node to
'free_ast_item'.

In main.c, replace usage of 'free_ast' with 'free_ast_item'.