refact(test): add test helper 'code2sexpr'
We were using command line flag '--sexpr' in tests to make onec only
parse source code and convert it into an S-expression. This was used to
test soundness of parser and AST nodes emitted by it. This flag however
is not something we would want to expose to the end-user as they might
not find value in it. We are therefore interested in moving S-expression
conversion somewhere else which is not shipped with onec.

In 1976fdf (feat(Makefile): add build targets for test helper programs)
and c67da19 (feat(tests): add test helper 'code2token-list'), we
introduced 't/helpers/' directory for housing C programs that assist in
testing a specific module of onec. We only had helper 'code2token-list'.

Lets make use of this infrastructure and introduce another helper
'code2sexpr' which invokes parser and S-expression printer.
In t/lib.py, replace the current implementation of 'code2sexpr()' that
invoked onec directly to instead call this helper.