fix(typecheck): ensure walk_function pops environment
When a function had a body, we were *not* popping the environment of
the function by the end, causing:

* The function environment to stay on top even after we have finished
  checking the function

* Memory leak.

This is because the function returned early when function node had a
body:

    environment_push(self->ctx->allocator, &self->ctx->env);

    ...

    if (fn->body != NULL) {
        return typecheck_stmt_list(self->ctx, fn->body);
    }

    environment_pop(self->ctx->allocator, &self->ctx->env);

Fix it by getting rid of the early return to ensure environment_pop is
called in call cases.

Add a simple test to verify that function environments stay isolated,
that is, a variable appearing in a previously type-checked function does
not slip into the function checked right after it, as this bug was
otherwise causing.