fix(alloc): free when requested new size is zero
In our allocator, we were relying on realloc to free the memory for us
when the requested new size is zero. This _works_ on glibc but might be
problematic elsewhere. Such use of realloc is now discouraged and is
undefined behaviour in C23.

Fix this by using free when new_size is 0. Our allocator's interface
does not change.