feat(parser): change parse()'s signature to support error recovery
The parse() function would return the AST when there are no syntax
errors and just die on encountering one.

We want the parse() to recover from errors and continue parsing, while
inserting "error nodes" when it is unable to build a proper AST due to
syntax errors. The callers should be notified if there was a syntax
error though.

Let's change parse()'s return type to bool - false indicating there was
at least one syntax error. We keep track of if an error was occurred in
`has_error` field of `parser_t`. Routines reporting errors bear the
responsibility of setting it to true. The AST is now written out to a
pointer provided as a parameter instead.

This does not yet change how the parser actually behaves. The parser
still dies on a syntax error, which will be addressed in subsequent
commits.