feat(parser): add a pass for validating curly braces
When doing a panic mode recovery from errors, in a statement
for example, parser must undergo "synchronization" where it discards
tokens until a delimiter/"statement boundary" is found. The parser can
continue parsing normally from there. Curly braces are such a token

(Another more common delimiter for statement is ';' but that does not
require such handling since they don't involve balancing with a closing
pair.)

However, when curly braces are not balanced in the source we are
parsing, the synchronization routine can end up in a rather weird
situation. Consider:

  fn main() {
    {
      let a = // ERROR
    }

    {
      let b = 0;

    // missing: '{'
  }

  fn other() {

  }