feat(parser): don't die on missing semi-colon
Since we have "non-dying" error reporting routines in place, let's use
it to report missing semi-colon without dying.

We can simply ignore a missing semi-colon because the line after a
statement is likely to be another statement (or end of block), which can
proceed normally without the semi-colon.

There might some funny cases though, consider:

  fn main() {
    let a = fun1() fun2();
  }

Here, we report a missing semi-colon after fun1(). However, it is likely
that the user is just missing an operator in between, and was trying to
write something like:

  let a = fun() + fun()

It is unfortunate that this gets reported with a 'missing semi-colon'
error, but there is no perfect.