fix(parser): tuple: pass 'function' as false when parsing a tuple
The code that parses function parameter types and tuples share a common
routine 'typename_tuple_items' because of the similarity in the syntax.
This routine takes in a boolean 'function' which tells us if we are
parsing a function or a tuple. We use this for error reporting.

When parsing a tuple, we were (incorrectly) passing in true for this
parameter which make the routine output inaccurate error messages.

Fix this by changing this argument to 'false'.