fix(parser): align '^^^' correctly when line contains tabs
We use '^^^' to point at the offending token when reporting errors.
In order to do align these characters with the token, we prefix it with it
with the same number of spaces as there are characters before the token:

    fn test offending() {
            ^^^^^^^^^

This works well when all the characters before the offending token are of
same width, however, when tabs are used in the same line before the token,
it is no longer aligned.

Fix this by inserting the same number of tabs as there are before the
offending token.