refact(parser): abstract error reporting from parser
In order to write tests that make assertions on the errors produced by
the parser, we have two choices:

1. Compare compiler's stderr with the expected error string.
   This isn't ideal since a change in an error message would require
   updating tests, even when we are merely fixing a grammatical error
   for example.

2. Make parser output errors in a different format; one that can be
   parsed from tests.

We take the second approach. In order for the parser to output errors in
a different format when it is invoked from tests, we need a way to swap
out error reporting implementation in parser.

Introduce error_printer: an indirection that parser calls to report
errors instead of printing them directly. Move the current error
printing implementation to error_printer and make this the default.
This way, the behaviour of error reporting does not see any real change
as far as the user is concerned.

This will be used in subsequent commits to make assertions on compiler
generated errors.