refact(lex): use error_printer to print lex errors
Since error_printer relied on tokens to point out the offending part of
the code, it was not possible for it to report errors generated by lexer
as we don't have any tokens we can work with when that happens.

Let's generalize error_printer to except a "span" in the source code,
which *may not* correspond to a valid token.

We extract fields that describe a token's span out of the token struct
so that we can use it in lex error struct just the same way. Lexical
errors now contain the span information which we can use to report
errors using error_printer the same way we report parse errors.

Hook lexical errors up to error_printer to make it all happen.