refact(cli): rid --typecheck-only flag
We have a CLI flag '--typecheck-only' which makes the compiler only run
typecheck and skip code execution. We used this in tests to make
assertions about typechecker.

Let's create a test helper similar to others in t/helpers/ that does
this for us, without having to bloat the CLI with this flag. Use this
helper to invoke typechecker and rid the flag.

We don't have any CLI flags anymore, but we still leave the argument
parsing code in place just in case we add any in future.