test: add integration test: 'file_compiles'
In the subsequent commit, we want to implemented memory mapped IO for
reading source files for parsing.

Before we do that, let's add a simple test that asserts that compiling a
file succeeds.

In order to do that, we must let the compiler exit with status 0 if the
typecheck passes. Currently we exit with 2 to indicate that code
execution/compilation is not implemented yet. Let's change that.